fix: move routeLeave code to NodeViewSwitcher
